requirements.txt
# requirements.txt

cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Deprecated==1.2.15
deprecation==2.1.0
img2pdf==0.5.1
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.1
ocrmypdf==16.7.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pi_heif==0.21.0
pikepdf==9.4.2
pillow==11.0.0
pluggy==1.5.0
pycparser==2.22
Pygments==2.18.0
PyMuPDF==1.25.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
six==1.17.0
tzdata==2024.2
wrapt==1.17.0
